+++ scripts/test/Console/Command/Dev/Tools/GenericInterface/DebugRead.t
@@ -1,0 +1,26 @@
+# --
+# Copyright (C) 2001-2016 OTRS AG, http://otrs.com/
+# --
+# This software comes with ABSOLUTELY NO WARRANTY. For details, see
+# the enclosed file COPYING for license information (AGPL). If you
+# did not receive this file, see http://www.gnu.org/licenses/agpl.txt.
+# --
+
+use strict;
+use warnings;
+use utf8;
+
+use vars (qw($Self));
+
+my $CommandObject = $Kernel::OM->Get('Kernel::System::Console::Command::Dev::Tools::GenericInterface::DebugRead');
+
+my $ExitCode = $CommandObject->Execute();
+
+# just check exit code
+$Self->Is(
+    $ExitCode,
+    0,
+    "Dev::Tools::ConsoleStats exit code",
+);
+
+1;